ChainRules API Documentation

ChainRules API Documentation

ChainRules.frule — Function.

frule(f, x...)

Expressing x as the tuple (x₁, x₂, ...) and the output tuple of f(x...) as Ω, return the tuple:

(Ω, (rule_for_ΔΩ₁::AbstractRule, rule_for_ΔΩ₂::AbstractRule, ...))

where each returned propagation rule rule_for_ΔΩᵢ can be invoked as

rule_for_ΔΩᵢ(Δx₁, Δx₂, ...)

to yield Ωᵢ's corresponding differential ΔΩᵢ. To illustrate, if all involved values are real-valued scalars, this differential can be written as:

ΔΩᵢ = ∂Ωᵢ_∂x₁ * Δx₁ + ∂Ωᵢ_∂x₂ * Δx₂ + ...

If no method matching frule(f, xs...) has been defined, then return nothing.

Examples:

unary input, unary output scalar function:

julia> x = rand();

julia> sinx, dsin = frule(sin, x);

julia> sinx == sin(x)
true

julia> dsin(1) == cos(x)
true

unary input, binary output scalar function:

julia> x = rand();

julia> sincosx, (dsin, dcos) = frule(sincos, x);

julia> sincosx == sincos(x)
true

julia> dsin(1) == cos(x)
true

julia> dcos(1) == -sin(x)
true

See also: rrule, AbstractRule, @scalar_rule

source

ChainRules.rrule — Function.

rrule(f, x...)

Expressing x as the tuple (x₁, x₂, ...) and the output tuple of f(x...) as Ω, return the tuple:

(Ω, (rule_for_Δx₁::AbstractRule, rule_for_Δx₂::AbstractRule, ...))

where each returned propagation rule rule_for_Δxᵢ can be invoked as

rule_for_Δxᵢ(ΔΩ₁, ΔΩ₂, ...)

to yield xᵢ's corresponding differential Δxᵢ. To illustrate, if all involved values are real-valued scalars, this differential can be written as:

Δxᵢ = ∂Ω₁_∂xᵢ * ΔΩ₁ + ∂Ω₂_∂xᵢ * ΔΩ₂ + ...

If no method matching rrule(f, xs...) has been defined, then return nothing.

Examples:

unary input, unary output scalar function:

julia> x = rand();

julia> sinx, dx = rrule(sin, x);

julia> sinx == sin(x)
true

julia> dx(1) == cos(x)
true

binary input, unary output scalar function:

julia> x, y = rand(2);

julia> hypotxy, (dx, dy) = rrule(hypot, x, y);

julia> hypotxy == hypot(x, y)
true

julia> dx(1) == (y / hypot(x, y))
true

julia> dy(1) == (x / hypot(x, y))
true

See also: frule, AbstractRule, @scalar_rule

source

ChainRules.AbstractRule — Type.

Subtypes of AbstractRule are types which represent the primitive derivative propagation "rules" that can be composed to implement forward- and reverse-mode automatic differentiation.

More specifically, a rule::AbstractRule is a callable Julia object generally obtained via calling frule or rrule. Such rules accept differential values as input, evaluate the chain rule using internally stored/ computed partial derivatives to produce a single differential value, then return that calculated differential value.

For example:

julia> using ChainRules: frule, rrule, AbstractRule

julia> x, y = rand(2);

julia> h, dh = frule(hypot, x, y);

julia> h == hypot(x, y)
true

julia> isa(dh, AbstractRule)
true

julia> Δx, Δy = rand(2);

julia> dh(Δx, Δy) == ((y / h) * Δx + (x / h) * Δy)
true

julia> h, (dx, dy) = rrule(hypot, x, y);

julia> h == hypot(x, y)
true

julia> isa(dx, AbstractRule) && isa(dy, AbstractRule)
true

julia> Δh = rand();

julia> dx(Δh) == (y / h) * Δh
true

julia> dy(Δh) == (x / h) * Δh
true

See also: frule, rrule, Rule, DNERule, WirtingerRule

source

ChainRules.Rule — Type.

Rule(propation_function[, updating_function])

Return a Rule that wraps the given propation_function. It is assumed that propation_function is a callable object whose arguments are differential values, and whose output is a single differential value calculated by applying internally stored/computed partial derivatives to the input differential values.

If an updating function is provided, it is assumed to have the signature u(Δ, xs...) and to store the result of the propagation function applied to the arguments xs into Δ in-place, returning Δ.

For example:

frule(::typeof(*), x, y) = x * y, Rule((Δx, Δy) -> Δx * y + x * Δy)

rrule(::typeof(*), x, y) = x * y, (Rule(ΔΩ -> ΔΩ * y'), Rule(ΔΩ -> x' * ΔΩ))

See also: frule, rrule, accumulate, accumulate!, store!

source

ChainRules.DNERule — Type.

DNERule(args...)

Construct a DNERule object, which is an AbstractRule that signifies that the current function is not differentiable with respect to a particular parameter. DNE is an abbreviation for Does Not Exist.

source

ChainRules.WirtingerRule — Type.

TODO

source

ChainRules.accumulate — Function.

accumulate(Δ, rule::AbstractRule, args...)

Return Δ + rule(args...) evaluated in a manner that supports ChainRules' various AbstractDifferential types.

This method intended to be customizable for specific rules/input types. For example, here is pseudocode to overload accumulate w.r.t. a specific forward differentiation rule for a given function f:

df(x) = # forward differentiation primitive implementation

frule(::typeof(f), x) = (f(x), Rule(df))

accumulate(Δ, rule::Rule{typeof(df)}, x) = # customized `accumulate` implementation

See also: accumulate!, store!, AbstractRule

source

ChainRules.accumulate! — Function.

accumulate!(Δ, rule::AbstractRule, args...)

Similar to accumulate, but compute Δ + rule(args...) in-place, storing the result in Δ.

Note that this function internally calls Base.Broadcast.materialize!(Δ, ...).

See also: accumulate, store!, AbstractRule

source

ChainRules.store! — Function.

store!(Δ, rule::AbstractRule, args...)

Compute rule(args...) and store the result in Δ, potentially avoiding intermediate temporary allocations that might be necessary for alternative approaches (e.g. copyto!(Δ, extern(rule(args...))))

Note that this function internally calls Base.Broadcast.materialize!(Δ, ...).

Like accumulate and accumulate!, this function is intended to be customizable for specific rules/input types.

See also: accumulate, accumulate!, AbstractRule

source

ChainRules.AbstractDifferential — Type.

The subtypes of AbstractDifferential define a custom "algebra" for chain rule evaluation that attempts to factor various features like complex derivative support, broadcast fusion, zero-elision, etc. into nicely separated parts.

All subtypes of AbstractDifferential implement the following operations:

add(a, b): linearly combine differential a and differential b

mul(a, b): multiply the differential a by the differential b

Base.conj(x): complex conjugate of the differential x

extern(x): convert x into an appropriate non-AbstractDifferential type for use outside of ChainContext.

Valid arguments to these operations are T where T<:AbstractDifferential, or where T has proper + and * implementations.

Additionally, all subtypes of AbstractDifferential support Base.iterate and Base.Broadcast.broadcastable(x).

source

ChainRules.extern — Function.

extern(x)

Return x converted to an appropriate non-AbstractDifferential type, for use with external packages that might not handle AbstractDifferential types.

Note that this function may return an alias (not necessarily a copy) to data wrapped by x, such that mutating extern(x) might mutate x itself.

source

ChainRules.Casted — Type.

TODO

source

ChainRules.Wirtinger — Type.

Wirtinger(primal::Union{Number,AbstractDifferential},
          conjugate::Union{Number,AbstractDifferential})

Returns a Wirtinger instance representing the complex differential:

df = ∂f/∂z * dz + ∂f/∂z̄ * dz̄

where primal corresponds to ∂f/∂z * dz and conjugate corresponds to ∂f/∂z̄ * dz̄.

The two fields of the returned instance can be accessed generically via the wirtinger_primal and wirtinger_conjugate methods.

source

ChainRules.Thunk — Type.

TODO

source

ChainRules.Zero — Type.

TODO

source

ChainRules.DNE — Type.

TODO

source

ChainRules.One — Type.

TODO

source